finished the transaction aspect of the system

+++ src/main/java/org/jiranibora/com/models/FineCategory.java
@@ -17,7 +17,7 @@ public class FineCategory {
     private String fineCategoryId;
     @Column(unique = true)
     private String fineName;
-    private Integer chargeableAmount;
+    private Double chargeableAmount;
 
 
 }
